DGM2221 - fixed reappearing feathers

+++ DGM2221 - Game Essentials/_Unity Files/PogoGame/Assets/Scripts/WeakenFlight.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class WeakenFlight : MonoBehaviour
 {
-    public IntData flightAmount;
+    public FloatData flightAmount;
     public int newFlightAmount;
 
     public void OnTriggerEnter(Collider other)
